pyproject.toml
[project]
name = "sentrysearch"
version = "0.1.0"
description = "Search dashcam footage using natural language queries"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "python-dotenv>=1.0",
    "chromadb>=0.4",
    "dashscope",
    "google-genai>=1.0",
    "requests>=2.31",
    "volcengine-python-sdk>=4.0.43",
    "imageio-ffmpeg>=0.5",
    "protobuf>=4.0,<7",
    "yt-dlp>=2025.3.31",
]

[project.optional-dependencies]
tesla = ["geopy"]
shots = ["transnetv2-pytorch>=1.0.5"]
local = [
    "torch>=2.0",
    "torchvision>=0.15,<0.22",
    "transformers>=5.3",
    "accelerate",
    "qwen-vl-utils",
    "torchcodec>=0.6",
]
local-quantized = [
    "torch>=2.0",
    "torchvision>=0.15,<0.22",
    "transformers>=5.3",
    "accelerate",
    "qwen-vl-utils",
    "torchcodec>=0.6",
    "bitsandbytes",
]

[project.scripts]
sentrysearch = "sentrysearch.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["sentrysearch"]

[dependency-groups]
test = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.uv]
constraint-dependencies = [
    "opentelemetry-exporter-otlp-proto-grpc>=1.25.0",
    "opentelemetry-proto>=1.25.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["sentrysearch"]
omit = ["sentrysearch/dashcam_pb2.py"]

[tool.coverage.report]
show_missing = true
